pyproject.toml
[project]
name = "consciousness"
version = "0.1.0"
description = "Self-referential AI daemon with continuous observation and decision-making"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "watchfiles>=0.21.0",
    "psutil>=5.9.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "structlog>=24.0.0",
    "aiosqlite>=0.20.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
consciousness = "consciousness.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["consciousness"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
